complex migration, database version 3

+++ app/src/main/kotlin/user/data/repository/UserRepository.kt
@@ -6,7 +6,7 @@ import user.data.datasource.room.UserRoomDataSource
 class UserRepository(private val userRoomDataSource: UserRoomDataSource) {
 
     fun insertUser(accountId: String, userName: String, email: String) =
-        userRoomDataSource.insert(UserEntity(accountId = accountId, userName = userName, email = email))
+        userRoomDataSource.insert(UserEntity(accountId = accountId.toInt(), userName = userName, email = email))
 
     fun getUsers() = userRoomDataSource.getAll()
 }
